backend/requirements.txt
# GenAI Investment Learning Platform - Backend Dependencies
# Phase 2

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Validation
pydantic>=2.5.0

# Environment Configuration
python-dotenv>=1.0.0

# HTTP Client for API calls
httpx>=0.25.0

# LLM Integration
openai>=1.3.0            # For GPT-4 integration
# anthropic>=0.7.0       # For Claude integration (optional)
# redis>=5.0.0           # For caching (future)
